+++ docs/12_0_0/gettingstarted/whatsnew.md
@@ -42,7 +42,7 @@ This page contains a list of big features. Please check the GitHub issues for al
   * Added `disableOnAjax` attribute to disable the button during Ajax requests triggered by it or its menu items.
   * Added Ajax load indicator (spinner) within the button during Ajax requests triggered by it.
 * TabMenu: added `orientation` (of the tab items relative to where you want to put the content) attribute.
-* TabView: added `ui-state-error` class to tabs that contain invalid inputs.
+* TabView: added `ui-state-error` class to tabs that contain invalid inputs and `focusOnError` if you want to focus the first tab with an error.
 * Tree: added `filterFunction` attribute for custom filtering.
 * ToggleSwitch: added `onIcon` and `offIcon` attributes
 * Tooltip: added `autoHide` attribute when set to false will keep the tooltip showing while focused.